SmartRecycle/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
# MySQL connector for database connection
mysql-connector-python==8.2.0
# PyMySQL==1.1.0
